+++ src/main/kotlin/edu/example/dev_3_5_cc/service/BoardService.kt
@@ -10,7 +10,9 @@ import edu.example.dev_3_5_cc.entity.Board
 import edu.example.dev_3_5_cc.entity.Category
 import edu.example.dev_3_5_cc.exception.AuthorizationException
 import edu.example.dev_3_5_cc.exception.BoardException
+import edu.example.dev_3_5_cc.exception.MemberException
 import edu.example.dev_3_5_cc.repository.BoardRepository
+import edu.example.dev_3_5_cc.repository.MemberRepository
 import edu.example.dev_3_5_cc.util.SecurityUtil
 import jakarta.transaction.Transactional
 import org.hibernate.query.sqm.tree.SqmNode.log
@@ -24,17 +26,20 @@ import org.springframework.stereotype.Service
 @Transactional
 class BoardService(
     private val boardRepository: BoardRepository,
-    private val securityUtil: SecurityUtil
+    private val securityUtil: SecurityUtil,
+    private val memberRepository: MemberRepository
 ) {
-
     fun createBoard (boardRequestDTO: BoardRequestDTO): BoardResponseDTO {
 
-        // 카테고리 권한 체크
-        checkCategoryAuthorization(boardRequestDTO.category)
+//        // 카테고리 권한 체크
+//        checkCategoryAuthorization(boardRequestDTO.category)
+//
+//        val currentUser = securityUtil.getCurrentUser()
+        val foundMember = memberRepository.findById(boardRequestDTO.memberId!!).orElseThrow {
+            MemberException.NOT_FOUND.get()
+        }
 
-        val currentUser = securityUtil.getCurrentUser()
-
-        val board = boardRequestDTO.toEntity(currentUser)
+        val board = boardRequestDTO.toEntity(foundMember!!)
 
         val savedBoard = boardRepository.save(board)
         return BoardResponseDTO(savedBoard)
@@ -48,17 +53,17 @@ class BoardService(
     fun updateBoard(boardUpdateDTO: BoardUpdateDTO): BoardResponseDTO {
         val board = boardRepository.findByIdOrNull(boardUpdateDTO.boardId) ?: throw BoardException.NOT_FOUND.get()
 
-        // 권한 체크: 작성자 또는 관리자만 가능
-        val member = board.member ?: throw AuthorizationException("이 글의 작성자 혹은 관리자만 가능합니다")
-        securityUtil.checkUserAuthorization(member)
-
-        // 카테고리 변경 검증: USER는 GENERAL -> TIP, NOTICE로 변경할 수 없음
-        val currentRole = securityUtil.getCurrentUserRole()
-        if ("ROLE_USER" == currentRole && board.category === Category.GENERAL &&
-            (boardUpdateDTO.category === Category.TIP || boardUpdateDTO.category === Category.NOTICE)
-        ) {
-            throw AuthorizationException("USER는 GENERAL 게시물을 TIP 또는 NOTICE로 변경할 수 없습니다.")
-        }
+//        // 권한 체크: 작성자 또는 관리자만 가능
+//        val member = board.member ?: throw AuthorizationException("이 글의 작성자 혹은 관리자만 가능합니다")
+//        securityUtil.checkUserAuthorization(member)
+//
+//        // 카테고리 변경 검증: USER는 GENERAL -> TIP, NOTICE로 변경할 수 없음
+//        val currentRole = securityUtil.getCurrentUserRole()
+//        if ("ROLE_USER" == currentRole && board.category === Category.GENERAL &&
+//            (boardUpdateDTO.category === Category.TIP || boardUpdateDTO.category === Category.NOTICE)
+//        ) {
+//            throw AuthorizationException("USER는 GENERAL 게시물을 TIP 또는 NOTICE로 변경할 수 없습니다.")
+//        }
 
         with(board){
             title = boardUpdateDTO.title
@@ -72,9 +77,9 @@ class BoardService(
     fun delete(boardId: Long){
         val board = boardRepository.findByIdOrNull(boardId) ?: throw BoardException.NOT_FOUND.get()
 
-        // 권한 체크: 작성자 또는 관리자만 가능
-        val member = board.member ?: throw AuthorizationException("이 글의 작성자 혹은 관리자만 가능합니다")
-        securityUtil.checkUserAuthorization(member)
+//        // 권한 체크: 작성자 또는 관리자만 가능
+//        val member = board.member ?: throw AuthorizationException("이 글의 작성자 혹은 관리자만 가능합니다")
+//        securityUtil.checkUserAuthorization(member)
 
         boardRepository.delete(board)
     }
@@ -93,13 +98,13 @@ class BoardService(
 
     fun listByMemberId(memberId : String) : List<BoardListDTO>{
 
-        // 현재 로그인한 사용자의 ID 가져오기
-        val currentUserId : String? = securityUtil.getCurrentUser().memberId
-
-        // 현재 사용자가 요청한 memberId와 동일한지 확인
-        if(memberId != currentUserId){
-            throw AuthorizationException("작성자만 자신의 게시물을 조회할 수 있습니다.")
-        }
+//        // 현재 로그인한 사용자의 ID 가져오기
+//        val currentUserId : String? = securityUtil.getCurrentUser().memberId
+//
+//        // 현재 사용자가 요청한 memberId와 동일한지 확인
+//        if(memberId != currentUserId){
+//            throw AuthorizationException("작성자만 자신의 게시물을 조회할 수 있습니다.")
+//        }
 
         val boards : List<Board> = boardRepository.findAllByMember(memberId)
         if (boards.isEmpty()) {
